+++ src/app/moved-media/tvshow/tvshow.component.ts
@@ -2,6 +2,8 @@ import {Component, OnInit} from '@angular/core';
 import {MovedMediaService, MovedMediaView} from "../moved-media.service";
 import {ActivatedRoute, Router} from "@angular/router";
 import {formatNumber} from "../moved-media.utils";
+import {MatDialog} from "@angular/material/dialog";
+import {ConfirmComponent} from "../confirm/confirm.component";
 
 @Component({
   selector: 'movedMedia-tvshow',
@@ -15,7 +17,8 @@ export class MovedMediaTVShowComponent implements OnInit {
 
     constructor(private route: ActivatedRoute,
                 private movedMediaService: MovedMediaService,
-                private router: Router) {}
+                private router: Router,
+                private dialog: MatDialog) {}
 
     ngOnInit() {
         this.route.params.subscribe(params => {
@@ -44,7 +47,13 @@ export class MovedMediaTVShowComponent implements OnInit {
     }
 
     async deleteShow() {
-        await this.movedMediaService.removeMovedMediaShow(this.mediaShowId!);
-        await this.router.navigate(['/moved-media/search']);
+        let config = { data: { questionKey: 'confirm delete' } };
+        const ref = this.dialog.open(ConfirmComponent, config);
+        ref.afterClosed().subscribe(async (result: boolean | undefined) => {
+            if (result === true) {
+                await this.movedMediaService.removeMovedMediaShow(this.mediaShowId!);
+                await this.router.navigate(['/moved-media/search']);
+            }
+        });
     }
 }
